hg_mitm_viewer.m: Include FPS in each GIF frame
The imwrite function defaults to an animated GIF inter-frame delay time of 0.5 seconds. It appears that imwrite expects the user to set the inter-frame time for each frame when writing. Without this addition, the resulting animated GIF would return to 2 FPS after the first frame was written. Including this extra input argument for each frame written allows for the hard-coded 4 FPS to be applied to every frame.